Improvements to the code and tests - added better package structure - added when the chat is greeted get a push notification - added initial tests but still disabled
