+++ pyproject.toml
@@ -4,6 +4,10 @@ version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.12"
+packages = [
+    {include = "src"}
+]
+
 dependencies = [
     "gradio>=5.22.0",
     "openai>=1.68.2",
@@ -11,7 +15,10 @@ dependencies = [
     "python-dotenv>=1.0.1",
     "faiss-cpu>=1.7.4",
     "numpy>=1.26.4",
+    "pytest"
 ]
+[tool.pytest]
+testpaths = ["tests"]
 
 [dependency-groups]
 dev = [
