+++ src/main/python/resources/summary.txt
